+++ Data/Mods/translation/PbFieldTranslation.rb
@@ -1,0 +1,78 @@
+def getFieldRootName(id)
+    case id
+    when PBFields::ELECTRICT
+        return "Electric Terrain"
+    when PBFields::GRASSYT
+        return "Grassy Terrain"
+    when PBFields::MISTYT
+        return "Misty Terrain"
+    when PBFields::DARKCRYSTALC
+        return "Dark Crystal Cavern"
+    when PBFields::CHESSB
+        return "Chess Board"
+    when PBFields::BIGTOPA
+        return "Big Top Arena"
+    when PBFields::BURNINGF
+        return "Burning Field"
+    when PBFields::SWAMPF
+        return "Swamp Field"
+    when PBFields::RAINBOWF
+        return "Rainbow Field"
+    when PBFields::CORROSIVEF
+        return "Corrosive Field"
+    when PBFields::CORROSIVEMISTF
+        return "Corrosive Mist Field"
+    when PBFields::DESERTF
+        return "Desert Field"
+    when PBFields::ICYF
+        return "Icy Field"
+    when PBFields::ROCKYF
+        return "Rocky Field"
+    when PBFields::FORESTF
+        return "Forest Field"
+    when PBFields::SUPERHEATEDF
+        return "Super-Heated Field"
+    when PBFields::FACTORYF
+        return "Factory Field"
+    when PBFields::SHORTCIRCUITF
+        return "Short-Circuit Field"
+    when PBFields::WASTELAND
+        return "Wasteland"
+    when PBFields::ASHENB
+        return "Ashen Beach"
+    when PBFields::WATERS
+        return "Water Surface"
+    when PBFields::UNDERWATER
+        return "Underwater"
+    when PBFields::CAVE
+        return "Cave"
+    when PBFields::GLITCHF
+        return "Glitch Field"
+    when PBFields::CRYSTALC
+        return "Crystal Cavern"
+    when PBFields::MURKWATERS
+        return "Murkwater Surface"
+    when PBFields::MOUNTAIN
+        return "Mountain"
+    when PBFields::SNOWYM
+        return "Snowy Mountain"
+    when PBFields::HOLYF
+        return "Holy Field"
+    when PBFields::MIRRORA
+        return "Mirror Arena"
+    when PBFields::FAIRYTALEF
+        return "Fairy Tale Field"
+    when PBFields::DRAGONSD
+        return "Dragon's Den"
+    when PBFields::FLOWERGARDENF
+        return "Flower Garden Field"
+    when PBFields::STARLIGHTA
+        return "Starlight Arena"
+    when PBFields::NEWW
+        return "New World"
+    when PBFields::INVERSEF
+        return "Inverse Field"
+    when PBFields::PSYCHICT
+        return "Psychic Terrain"
+    end
+end
