+++ Scripts/PokemonPokegear.rb
@@ -103,7 +103,7 @@ class Scene_Pokegear
     @button=AnimatedBitmap.new("Graphics/Pictures/Pokegear/pokegearButton")
     @sprites={}
     @sprites["background"] = IconSprite.new(0,0)
-    @sprites["background"].setBitmap("Graphics/Pictures/Pokegear/pokegearbg")
+    @sprites["background"].setBitmap("Graphics/Pictures/Pokegear/pokegearbg" + getSuffixFile())
     @sprites["command_window"] = Window_CommandPokemon.new(commands,160)
     @sprites["command_window"].index = @menu_index
     @sprites["command_window"].x = Graphics.width
