+++ Scripts/PokemonStorage.rb
@@ -1855,7 +1855,7 @@ class PokemonBoxPartySprite < SpriteWrapper
 
   def initialize(party,viewport=nil)
     super(viewport)
-    @boxbitmap=AnimatedBitmap.new("Graphics/Pictures/Storage/boxpartytab")
+    @boxbitmap=AnimatedBitmap.new("Graphics/Pictures/Storage/boxpartytab" + getSuffixFile())
     @pokemonsprites=[]
     @party=party
     for i in 0...6
@@ -2170,7 +2170,7 @@ class PokemonStorageScene
     addBackgroundPlane(@sprites,"background","boxbg",@bgviewport)
     @sprites["box"]=PokemonBoxSprite.new(@storage,@storage.currentBox,@boxviewport)
     @sprites["boxsides"]=IconSprite.new(0,0,@boxsidesviewport)
-    @sprites["boxsides"].setBitmap("Graphics/Pictures/Storage/boxsides")
+    @sprites["boxsides"].setBitmap("Graphics/Pictures/Storage/boxsides" + getSuffixFile())
     @sprites["overlay"]=BitmapSprite.new(Graphics.width,Graphics.height,@boxsidesviewport)
     @sprites["pokemon"]=AutoMosaicPokemonSprite.new(@boxsidesviewport)
     pbSetSystemFont(@sprites["overlay"].bitmap)
