+++ Scripts/Reborn/RebornScripts.rb
@@ -34,13 +34,23 @@ def pbTicketText(textno)
   end    
   baseColor=Color.new(78,66,66)
   shadowColor=Color.new(159,150,144) 
+  onePosition = 60
+  grandviewPosition = 73
+  trainPlacePosition = 98
+  numberPosition = 83
+  if (LANGUAGES.length>=2 && LANGUAGES[$idk[:settings].language][1] == "fr")
+    onePosition = 50
+    grandviewPosition = 74
+    trainPlacePosition = 102
+    numberPosition = 85
+  end
   textPositions=[
       [playerName,(Graphics.width/2)-143,32+166,0,baseColor,shadowColor],       
       [playerGender,(Graphics.width/2)+26,32+166,0,baseColor,shadowColor],       
-      ["8R750",(Graphics.width/2)-83,32+189,0,baseColor,shadowColor],       
-      ["5D",(Graphics.width/2)+98,32+189,0,baseColor,shadowColor],       
-      ["Gare Bellevue",(Graphics.width/2)-73,32+216,0,baseColor,shadowColor],       
-      ["UN",(Graphics.width/2)-60,32+241,0,baseColor,shadowColor],       
+      ["8R750",(Graphics.width/2)-numberPosition,32+189,0,baseColor,shadowColor],       
+      ["5D",(Graphics.width/2)+trainPlacePosition,32+189,0,baseColor,shadowColor],       
+      [_INTL("Grandview Station"),(Graphics.width/2)-grandviewPosition,32+216,0,baseColor,shadowColor],       
+      [_INTL("ONE"),(Graphics.width/2)-onePosition,32+241,0,baseColor,shadowColor],       
       ["SGL",(Graphics.width/2)+98,32+241,0,baseColor,shadowColor],       
   ]
   finalTextPositions=[textPositions[textno]]
