+++ Scripts/idk.rb
@@ -56,7 +56,7 @@ def startup
     pbSetUpSystem
     
     
-    Dir["./Data/Mods/*.rb"].each {|file| load File.expand_path(file) }
+    Dir["./Data/Mods/**/*.rb"].each {|file| load File.expand_path(file) }
     $cache.trainers = load_data("Data/trainers" + getSuffixFile() + ".dat") if !@trainers
 end
