--- errorTranslation.txt
@@ -1,157 +1,0 @@
-547 : "Fairview 34" 
-547 : "Fairview 26" 
-547 : "Fairview 22" 
-553 : "Fairview 14" 
-553 : "Fairview 1" 
-553 : "Fairview 10" 
-547 : "Fairview 19" 
-547 : "Fairview 29" 
-553 : "Sunrise 28" 
-553 : "Sunrise 20" 
-553 : "Sunrise 3" 
-553 : "Sunrise 4" 
-553 : "Sunrise 5" 
-553 : "Sunrise 9" 
-553 : "Sunrise 12" 
-558 : "Kingsbury 8" 
-558 : "Kingsbury 17" 
-558 : "Kingsbury 22" 
-558 : "Kingsbury 28" 
-558 : "Kingsbury 33" 
-558 : "Kingsbury 25" 
-558 : "Kingsbury 19" 
-574 : "Kingsbury 5" 
-574 : "Kingsbury 8" 
-574 : "Kingsbury 2" 
-38 : Not bad. Or maybe I've just lost my touch.
-
-552 : "Lyssa Copperman" 
-25 : 'Ey kid! You want me to teach your Pokémon some special moves? Take a look what I got! 
-552 : "It's been a while." 
-552 : "Did you find the peace you were looking for?" 
-764 : "Euphie" 
-372 : Buddy, I really don't got anything else for you, we're done here. 
-498 : "G.U.M. Room" 
-606 : "ACTIVATING SYSTEM..." 
-757 : "Defeat the inner chaos?" 
-757 : "Look! 
-758 : Noooooh...
-413 : Many high ranking Meteors have treated their Pokémon with PULSE2s by now. 
-753 : "I didn't know..." 
-753 : "He tricked me..." 
-753 : "I have to fix it..." 
-753 : "I'll get them all back..." 
-413 : I got tired of being tired, so I thought I'd come up here and say "hi"! 
-413 : HARDY: Whoa there, you don't have to hit the guy. 
-209 : "how dare u read me" 
-414 : "PULSE-10" 
-414 : "No, please, stop." 
-414 : "Stop, stop--" 
-775 : "PULSE-09" 
-769 : #carsfuckingdragons #feminism 
-776 : "Y0u almost done in there, h0ney?" m0aned Lt. Surge, 
-776 : "I need to... relieve mys3lf..........." 
-770 : my whole life......... has led up to this moment....... 
-770 : 'preciate u bbg 
-771 : N0W HOLD ON JUST SEXOND!!! 
-771 : \PN sensssualllly RIPPED OFF HIS UNIFORM, REVEALLING HIS SMOOTH, LANKY B0DY B3FOrE ape-aggressing n2 his SHOWAA. 
-191 : "Heather _____" 
-191 : "Noel Doe" 
-191 : "Laura Belrose" 
-191 : "Saphira Belrose" 
-191 : AGE AT INTAKE: 15 years 
-191 : "Lindsey Paraeah" 
-191 : "---phie Villeneuvelle Myt-----" 
-190 : "Lower-Level Arena Access" 
-190 : "Boys' Room" 
-190 : "Girls' Room" 
-780 : "Lizzy, Lizzy, Lizzy," 
-785 : "SIZE:" 
-785 : "Large" 
-785 : "Miniscule" 
-785 : "Small" 
-785 : "HARDNESS:" 
-785 : "5" 
-785 : "7" 
-785 : "PURITY:" 
-785 : "Impure" 
-785 : "Middling" 
-785 : "6" 
-785 : "Medium" 
-785 : "Pure" 
-785 : "Flawless" 
-785 : "8" 
-789 : "SIZE:" 
-789 : "HARDNESS:" 
-789 : "PURITY:" 
-789 : "FOLIATION:" 
-789 : "Perfect" 
-789 : "Indistinct" 
-789 : "Difficult" 
-789 : "Eminent" 
-789 : "Small" 
-789 : "Miniscule" 
-789 : "7" 
-789 : "Flawless" 
-789 : "Pure" 
-789 : "Impure" 
-789 : "Middling" 
-789 : "6" 
-793 : "SIZE:" 
-793 : "HARDNESS:" 
-793 : "PURITY:" 
-793 : "FOLIATION:" 
-793 : "LUSTER:" 
-793 : "Silky" 
-793 : "Vitreous" 
-793 : "Pearly" 
-793 : "Adamantine" 
-793 : "Eminent" 
-793 : "Difficult" 
-793 : "Indistinct" 
-793 : "Perfect" 
-793 : "Pure" 
-793 : "Miniscule" 
-793 : "Middling" 
-793 : "Small" 
-793 : "Medium" 
-793 : "5" 
-793 : "Impure" 
-795 : "HABIT:" 
-795 : "Hexagonal" 
-795 : "Cubic" 
-795 : "Tabular" 
-795 : "Coxcomb" 
-785 : "Complete." 
-808 : "Wait in the crystal branches," she said. 
-808 : "Just buy more time," she said. 
-796 : "Crystal Star" items are inserted into the device's input via the slots below. 
-798 : "What right do you have?" you might ask. 
-713 : \PN received 5 PokéSnax! 
-38 : Teenage Mutant Ninja Squirtles!!!
-0 : jshdkjhfd
-824 : "Absolute corruption must always be answered and blah blah blah--" 
-827 : Please note that we currently have a special offer on the EXP All Upgrade! 
-827 : If you purchase it, you will be elligible to purchase another special item as well-- the Remote PC. 
-827 : NIHIL: I wondered if I'd see you around here. 
-827 : They got me on security detail. 
-827 : Just stand watch, and toss out the loud people. 
-827 : Doesn't happen often. 
-827 : Good vibes. 
-827 : You sure this qualifies as "honest work"? 
-827 : Whatever. 
-827 : I mean, thanks. 
-827 : I mean... whatever. 
-827 : "That's nonsense," I told her, 
-830 : INSPECTOR: Don't be a stranger. 
-39 : -RANDOM STARTER SERVICE- 
-71 : "XX7X" 
-71 : "X1XX" 
-71 : "3XXX" 
-71 : "XXX9" 
-71 : "XX6X" 
-71 : "XXX2" 
-71 : "X8XX" 
-71 : "4XXX" 
-71 : "SECONDARY SECURITY" 
-38 : Hmm. That was quite interesting.
